Auth

Required state

class (MonadIO m, FilterMonad Response m, ServerMonad m, HasRqData m, Functor m) => AuthMonad m where

Methods

getAuthState :: m AuthState

data AuthState

This type is used to customize how authentication works for your site. We recommend building it from the defaultAuthState value.

Constructors

MkAuthState

Fields

loginForm :: String

where to redirect to to handle logins

loginSecret :: Maybe Key

key to use for encryption

loginCookieName :: Maybe String

Name of cookie to store authentication data

defaultAuthState :: AuthState

Key management

getKey :: FilePath -> IO Key

defaultKeyFile :: FilePath

getDefaultKey :: IO Key

initKey :: ByteString -> Either String Key

System hooks, login/logout

setLoggedIn :: (SafeCopy user, AuthMonad m) => user -> m ()

NOTE: the passed in user-data is sent to the client encrypted. Call this in the handler for your login form.

forwardAfterLogin

Arguments

:: AuthMonad m

=> String
default redirect url

-> m ()

Return a redirect to send the user back where they were trying to go when they were bounced to the login form. NOTE: You'll lose any post-body data the user was trying to submit, so try to require login on the form GET as well as the form POST.

loginData :: (SafeCopy u, AuthMonad m) => m (Maybe u)

If there is a logged in user and the session data is valid returned the session data.

requireLoggedIn :: (SafeCopy user, AuthMonad m) => m user

Return the logged-in user. If a user is not logged in, they are forwarded to your login page. After the

setLoggedOut :: m ()

If a user is logged in, log them out. We do not gaurantee this succeeds for a malicious user agent.

refreshLoggedIn :: m ()

If a user is logged in, refresh their login cookie so they can stay logged in. Used for keeping a user active when you want to log them out after X minutes for inactivity.